Update README.md
This python project basically, tic tac toe is 2 player game that is played in a 3x3 matrix/board. The player 1 chooses a marker X or O and other player chooses the remaining one. So now the game play, player 1 then followed by player 2 will play in the order, each player will place their respective markers on the board, in the sequence. the game ends by either one player wins or match ties. The winning position will be 3 marker of a player in horizontal, vertical or diagonal way.it will be like 3+3+2 possibilities, else the match is consider to be tie.
So in normal game play, the game is played between 2 player but in my project 1 player can play against the computer based on a simple algorithm, and it’s the simple program so no big presenting tools were used.